Use appropriate product name in inspections
There are two official names: "Magento Open Source" and "Adobe Commerce".
As the product "Magento Open Source" is the foundation of "Adobe Commerce"
because "Adobe Commerce" can not be used without "Magento Open Source",
and since the plugin works for either, using "Magento Open Source" is a
more appropriate name for anything that applies to both
"Magento Open Source" and "Adobe Commerce".

The name "Adobe Commerce" is appropriate for anything that is exclusive to
 "Adobe Commerce" and does not apply to "Magento Open Source".

Describe alternatives you've considered:

An alternative would be to always explicitly name both products, as in
"Magento Open Source or Adobe Commerce" but that is excessively long".

Additional context:

Personally I think it would be better to only use the name "Magento"
when referring to both products, but probably that won't fly with Adobe.